+++ app/Http/Controllers/RoleController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Entries;
 use App\Models\Role;
+use App\Models\SuccessIndicator;
+use App\Models\User;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Crypt;
@@ -17,7 +20,45 @@ class RoleController extends Controller
     public function roles()
     {
         $user=Auth::user();
-        return view('role_page.roles', compact('user'));
+
+        $currentYear = Carbon::now()->format('Y');
+        $currentUser = Auth::user();
+        $entriesCount = SuccessIndicator::whereNull('deleted_at')->whereYear('created_at', $currentYear);
+
+        $indicators = $entriesCount->get();
+        
+        $userDivisionIds = json_decode($currentUser->division_id, true);
+        $filteredIndicators = $indicators->filter(function($indicator) use ($userDivisionIds) {
+            $indicatorDivisionIds = json_decode($indicator->division_id, true);
+            
+            return !empty(array_intersect($userDivisionIds, $indicatorDivisionIds));
+        });
+
+        $currentMonth = Carbon::now()->format('m');
+        $current_Year = Carbon::now()->format('Y');
+
+        $currentDate = Carbon::now();
+
+        if ($currentDate->day > 5) {
+            $targetMonth = $currentDate->month;
+            // $targetMonth = $currentDate->addMonth()->month;
+        } else {
+            $targetMonth = $currentDate->subMonth()->month;
+        }
+
+        $filteredIndicators = $filteredIndicators->filter(function($indicator) use ($targetMonth, $current_Year) {
+            $completedEntries = Entries::where('indicator_id', $indicator->id)
+                                    ->whereMonth('created_at', $targetMonth)
+                                    ->whereYear('created_at', $current_Year)
+                                    ->where('status', 'Completed')
+                                    ->where('user_id',  Auth::user()->id)
+                                    ->exists();
+            return !$completedEntries;
+        });
+          
+            // $entriesCount = Entries::whereNull('deleted_at')->with('indicator')->where('status', 'Pending')->count();
+        $entriesCount = $filteredIndicators->count();
+        return view('role_page.roles', compact('user', 'entriesCount'));
     }
 
     public function getRole(Request $request){
@@ -125,7 +166,15 @@ class RoleController extends Controller
             return response()->json(['success' => false, 'errors' => $validator->errors()], 200);
         }
 
+        $ifExist = User::whereNull('deleted_at')->where('role_id', Crypt::decrypt($request->id))->exists();
+
+        if($ifExist){
+            
+            return response()->json(['success' => false, 'errors' => 'The role is being use in by users, Cannot be deleted']);
+        }
+
         $role = Role::findOrFail(Crypt::decrypt($request->id));
+        $role->created_by = Auth::user()->user_name;
         $role->delete();
 
         return response()->json(['success' => true, 'message' => 'Role deleted successfully']);
